# packages/server/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Nestx-lte","description":"Nestx-lte REST API provider.","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"KeyValue":{"type":"object","properties":{"label":{"type":"string"},"value":{"type":"string"}},"required":["label","value"]},"ErrorResult":{"type":"object","properties":{"code":{"type":"number"},"message":{"type":"string"}},"required":["code","message"]},"CreateAccountDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"keyword":{"type":"string"},"avatar":{"type":"string"},"type":{"type":"string"},"groups":{"type":"array","items":{"type":"string"}},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"},"mobile":{"type":"string"},"profile":{"type":"string"},"isDisable":{"type":"boolean"},"isAdmin":{"type":"boolean"},"isApproved":{"type":"boolean"},"secret":{"type":"string"},"expired":{"type":"number"}},"required":["username","password","name","keyword","avatar","type","groups","roles","email","mobile","profile","isDisable","isAdmin","isApproved","secret","expired"]},"AccountDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"keyword":{"type":"string"},"avatar":{"type":"string"},"type":{"type":"string"},"groups":{"type":"array","items":{"type":"string"}},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"},"mobile":{"type":"string"},"profile":{"type":"string"},"isDisable":{"type":"boolean"},"isAdmin":{"type":"boolean"},"isApproved":{"type":"boolean"},"secret":{"type":"string"},"expired":{"type":"number"}},"required":["id","username","password","name","keyword","avatar","type","groups","roles","email","mobile","profile","isDisable","isAdmin","isApproved","secret","expired"]},"EditAccountDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"keyword":{"type":"string"},"avatar":{"type":"string"},"type":{"type":"string"},"groups":{"type":"array","items":{"type":"string"}},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"},"mobile":{"type":"string"},"profile":{"type":"string"},"isDisable":{"type":"boolean"},"isAdmin":{"type":"boolean"},"isApproved":{"type":"boolean"},"secret":{"type":"string"},"expired":{"type":"number"}},"required":["id","username","password","name","keyword","avatar","type","groups","roles","email","mobile","profile","isDisable","isAdmin","isApproved","secret","expired"]},"AccountItemDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"keyword":{"type":"string"},"avatar":{"type":"string"},"type":{"type":"string"},"groups":{"type":"array","items":{"type":"string"}},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"},"mobile":{"type":"string"},"profile":{"type":"string"},"isDisable":{"type":"boolean"},"isAdmin":{"type":"boolean"},"isApproved":{"type":"boolean"},"secret":{"type":"string"},"expired":{"type":"number"}},"required":["id","username","password","name","keyword","avatar","type","groups","roles","email","mobile","profile","isDisable","isAdmin","isApproved","secret","expired"]},"AccountPaginateDto":{"type":"object","properties":{"page":{"type":"number"},"size":{"type":"number"},"count":{"type":"number"},"items":{"type":"array","items":{"$ref":"#/components/schemas/AccountItemDto"}}},"required":["page","size","count","items"]}}},"paths":{"/rest/core/account/search":{"get":{"operationId":"accountSearch","summary":"Account Search","parameters":[{"name":"keyword","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/KeyValue"}}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResult"}}}}},"tags":["Account"]}},"/rest/core/account/create":{"post":{"operationId":"accountCreate","summary":"Account Create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDto"}}}}},"tags":["Account"]}},"/rest/core/account/update":{"put":{"operationId":"accountUpdate","summary":"Account Update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditAccountDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDto"}}}}},"tags":["Account"]}},"/rest/core/account/query":{"get":{"operationId":"accountQuery","summary":"Account Query","parameters":[{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}},{"name":"size","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"undefined":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountPaginateDto"}}}}},"tags":["Account"]}},"/rest/core/account/delete/{id}":{"delete":{"operationId":"accountDelete","summary":"Account Delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Account"]}},"/rest/core/account/find-one/{id}":{"get":{"operationId":"accountFindOne","summary":"Account Detail","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDto"}}}}},"tags":["Account"]}}}}